// apps/mobile/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5"
  }
  public: {
    Tables: {
// ...
      records: {
        Row: {
          category_id: number | null
          created_at: string
          date: string
          description: string | null
          duration: number | null
          end_time: string | null
          id: number
          is_user_category: boolean
          start_time: string | null
          user_category_id: number | null
          user_id: string
        }
        Insert: {
          category_id?: number | null
          created_at?: string
          date: string
          description?: string | null
          duration?: number | null
          end_time?: string | null
          id?: number
          is_user_category?: boolean
          start_time?: string | null
          user_category_id?: number | null
          user_id?: string
        }
        Update: {
          category_id?: number | null
          created_at?: string
          date?: string
          description?: string | null
          duration?: number | null
          end_time?: string | null
          id?: number
          is_user_category?: boolean
          start_time?: string | null
          user_category_id?: number | null
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "records_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "records_user_category_id_fkey"
            columns: ["user_category_id"]
            isOneToOne: false
            referencedRelation: "user_categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      user_categories: {
        Row: {
          color: string
          created_at: string
          id: number
          subtitle: string
          title: string
          user_id: string
          workload: number
        }
        Insert: {
          color: string
          created_at?: string
          id?: number
          subtitle: string
          title: string
          user_id: string
          workload: number
        }
        Update: {
          color?: string
          created_at?: string
          id?: number
          subtitle?: string
          title?: string
          user_id?: string
          workload?: number
        }
        Relationships: []
      }
// ...
      users: {
        Row: {
          canton_code: string
          class_size: number | null
          created_at: string
          custom_work_hours: number
          education_level: Database["public"]["Enums"]["education_level"] | null
          first_name: string
          is_mode_dark: boolean
          language: string
          last_name: string
          stat_end_date: string | null
          stat_start_date: string | null
          teacher_relief: number | null
          user_id: string
          workload: number | null
        }
        Insert: {
          canton_code: string
          class_size?: number | null
          created_at?: string
          custom_work_hours?: number
          education_level?:
            | Database["public"]["Enums"]["education_level"]
            | null
          first_name: string
          is_mode_dark: boolean
          language?: string
          last_name: string
          stat_end_date?: string | null
          stat_start_date?: string | null
          teacher_relief?: number | null
          user_id: string
          workload?: number | null
        }
        Update: {
          canton_code?: string
          class_size?: number | null
          created_at?: string
          custom_work_hours?: number
          education_level?:
            | Database["public"]["Enums"]["education_level"]
            | null
          first_name?: string
          is_mode_dark?: boolean
          language?: string
          last_name?: string
          stat_end_date?: string | null
          stat_start_date?: string | null
          teacher_relief?: number | null
          user_id?: string
          workload?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "users_canton_code_fkey"
            columns: ["canton_code"]
            isOneToOne: false
            referencedRelation: "cantons"
            referencedColumns: ["code"]
          },
          {
            foreignKeyName: "users_language_fkey"
            columns: ["language"]
            isOneToOne: false
            referencedRelation: "languages"
            referencedColumns: ["locale"]
          },
        ]
      }
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never